Don't apply CMK encryption to public buckets
If a bucket is encrypted with customer-managed-key encryption,
its files cannot be served publicly because they can only be
accessed using properly authorized API calls.

This change notices if encryption is enabled generally and customer
managed keys are also enabled, and just for the public assets
bucket, sets its encryption to AES256 instead.